requirements.txt
redis==4.5.5
Django==4.2.1
psycopg==3.1.9
python-decouple==3.8
django-cors-headers==4.0.0
djangorestframework==3.14.0
